pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "esmecata"
dynamic = ["version"]
authors = [{name = "AuReMe", email = "gem-aureme@inria.fr"}]
readme = "README.md"
description = "EsMeCaTa: Estimating Metabolic Capabilties from Taxonomy"
license = {text = "GPL-3.0-or-later"}

dependencies = [
  'biopython',
  'bioservices',
  'matplotlib',
  'pandas',
  'requests',
  'ete3',
  'seaborn',
  'SPARQLWrapper'
]

[project.scripts]
esmecata = "esmecata.__main__:main"
esmecata_report = "esmecata.__main_report__:main"
esmecata_gseapy = "esmecata.__main_gseapy__:main"
esmecata_create_db = "esmecata.__main_create_database__:main"

[project.urls]
Homepage = "https://github.com/AuReMe/esmecata"
Changelog = "https://github.com/AuReMe/esmecata/blob/main/CHANGELOG.md"

[project.optional-dependencies]
ANALYSIS = ['datapane',
  'plotly',
  'kaleido',
  'ontosunburst'
]
GSEAPY = ['gseapy',
  'orsum',
  'pronto'
]
TEST = ['pytest',
  'pytest-mock'
]

[tool.setuptools]
packages = ['esmecata', 'esmecata.core', 'esmecata.report', 'esmecata.gseapy']
package-dir = {'esmecata' = 'esmecata', 'esmecata.core' = 'esmecata/core' , 'esmecata.report' = 'esmecata/report', 'esmecata.gseapy' = 'esmecata/gseapy'}

[tool.setuptools.dynamic]
version = { attr = "esmecata.__version__" }
dependencies = {file = ["requirements.txt"]}
